# api/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Grocery & Meal Planner API
  description: |
    MVP API for managing pantry inventory, grocery lists, meal ideas, and expiration notifications.
    
    ## Features
    - **Inventory Management**: CRUD operations for pantry items with low-stock and expiring-soon flags
    - **Grocery Lists**: Create and manage shopping lists, auto-add low stock or expiring items
    - **Meal Ideas**: Get suggestions based on pantry contents, save favorite meals
    - **Notifications**: Schedule reminders for expiring items
  version: 1.0.0
  contact:
    name: Smart Grocery Team
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/pantry:
    get:
      tags: [Pantry]
      summary: List pantry items for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of pantry items with computed flags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PantryItemListResponse'
    post:
      tags: [Pantry]
      summary: Create pantry item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePantryItem'
      responses:
        '201':
          description: Item created
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/pantry/expiring:
    get:
      tags: [Pantry]
      summary: Get items expiring soon
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        '200':
          description: Expiring items
  /api/pantry/low-stock:
    get:
      tags: [Pantry]
      summary: Get low stock items
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Low stock items
  /api/pantry/categories:
    get:
      tags: [Pantry]
      summary: Get unique categories
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of categories
  /api/pantry/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Pantry]
      summary: Get pantry item
      responses:
        '200':
          description: Pantry item
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Pantry]
      summary: Update pantry item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePantryItem'
      responses:
        '200':
          description: Item updated
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Pantry]
      summary: Delete pantry item
      responses:
        '200':
          description: Item deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # Grocery Lists
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: List notifications for user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, sent, cancelled]
      responses:
        '200':
          description: List of notifications
    post:
      tags: [Notifications]
      summary: Create notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNotification'
      responses:
        '201':
          description: Notification created
  /api/notifications/pending:
    get:
      tags: [Notifications]
      summary: Get pending notifications that are due
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Due notifications
  /api/notifications/auto-schedule:
    post:
      tags: [Notifications]
      summary: Auto-schedule notifications for expiring items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: integer
      responses:
        '200':
          description: Notifications scheduled
  /api/notifications/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Notifications]
      summary: Get notification
      responses:
        '200':
          description: Notification
    put:
      tags: [Notifications]
      summary: Update notification
      responses:
        '200':
          description: Notification updated
    delete:
      tags: [Notifications]
      summary: Delete notification
      responses:
        '200':
          description: Notification deleted
  /api/notifications/{id}/mark-sent:
    put:
      tags: [Notifications]
      summary: Mark notification as sent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Marked as sent
  /api/notifications/{id}/cancel:
    put:
      tags: [Notifications]
      summary: Cancel notification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Notification cancelled
components:
  schemas:
...
    CreatePantryItem:
      type: object
      required: [userId, name]
      properties:
        userId:
          type: integer
        name:
          type: string
          maxLength: 255
        quantity:
          type: number
        unit:
          type: string
          maxLength: 50
        category:
          type: string
          maxLength: 100
        expirationDate:
          type: string
          format: date
        source:
          type: string
          maxLength: 100
    UpdatePantryItem:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        category:
          type: string
        expirationDate:
          type: string
          format: date
    PantryItem:
      type: object
      properties:
        pantryItemId:
          type: integer
        userId:
          type: integer
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        category:
          type: string
        expirationDate:
          type: string
          format: date
        source:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isExpiringSoon:
          type: boolean
        isLowStock:
          type: boolean
        daysUntilExpiry:
          type: integer
    PantryItemListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/PantryItem'

    # Grocery List schemas
...
    CreateNotification:
      type: object
      required: [pantryItemId, scheduledFor]
      properties:
        pantryItemId:
          type: integer
        scheduledFor:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, sent, cancelled]
          default: pending
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  message:
                    type: string
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object